Widen heartbeat/worker-state PKs to BIGINT (schema v4)
horsies_heartbeats.id and horsies_worker_states.id were int4
autoincrement: at one runner heartbeat per RUNNING task per 30s plus
claimer beats, ~3000 in-flight tasks exhaust the sequence in months of
uptime. horsies_task_attempts already used BIGSERIAL — these two were
the outliers. Models use BigInteger; migration ALTERs existing columns
(both tables are retention-bounded, so the rewrite is cheap).

+++ horsies/core/models/task_pg.py
@@ -229,7 +229,7 @@ class TaskHeartbeatModel(Base):
         ),
     )
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
     task_id: Mapped[str] = mapped_column(String(36), nullable=False)
     sender_id: Mapped[str] = mapped_column(String(255), nullable=False)
     role: Mapped[str] = mapped_column(String(20), nullable=False)
@@ -252,7 +252,7 @@ class WorkerStateModel(Base):
     __tablename__ = 'horsies_worker_states'
 
     # Timeseries primary key
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
     worker_id: Mapped[str] = mapped_column(String(255), nullable=False, index=True)
     snapshot_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True),
